webapps2023/requirements.txt
Django==4.2
djangorestframework==3.14.0
pip==22.3.1
wheel==0.38.4
cryptography==40.0.2
pytz==2023.3
setuptools==65.5.1
asgiref==3.6.0
sqlparse==0.4.4
requests==2.29.0
certifi==2022.12.7
cffi==1.15.1
Markdown==3.4.3
idna==3.4
urllib3==1.26.15
MarkupSafe==2.1.2
Werkzeug==2.3.3
pycparser==2.21
zipp==3.15.0
django-crispy-forms==2.0
django-filter==23.1